+++ src/test/kotlin/com/asyncapi/plugin/idea/_core/AsyncAPISchemaRecognizerTest.kt
@@ -1,13 +1,12 @@
 package com.asyncapi.plugin.idea._core
 
-import com.intellij.json.psi.JsonFile
 import com.intellij.lang.Language
 import com.intellij.openapi.components.service
+import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import junit.framework.TestCase
 import org.jetbrains.yaml.YAMLLanguage
-import org.jetbrains.yaml.psi.YAMLFile
 
 /**
  * @author Pavel Bodiachevskii
@@ -32,56 +31,179 @@ class AsyncAPISchemaRecognizerTest: BasePlatformTestCase() {
         )
     }
 
-    fun `test - isSchema(psiFile) when file is schema and not empty`() {
+    fun `test - extractAsyncAPIKey when file is null`() {
         val asyncAPISchemaRecognizer = service<AsyncAPISchemaRecognizer>()
 
-        val psiFileFactory = PsiFileFactory.getInstance(project)
+        TestCase.assertNull(
+            "must be false in case of null file.",
+            asyncAPISchemaRecognizer.extractAsyncAPIKey(null)
+        )
+    }
+
+    fun `test - isSupported when file is null`() {
+        val asyncAPISchemaRecognizer = service<AsyncAPISchemaRecognizer>()
+
+        TestCase.assertFalse(
+            "must be false in case of null version.",
+            asyncAPISchemaRecognizer.isSupported(null)
+        )
+    }
+
+    fun `test (json) - 2_0_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.0.0.json", Language.findLanguageByID("JSON")!!),
+            "2.0.0"
+        )
+    }
+
+    fun `test (yaml) - 2_0_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.0.0.yaml", YAMLLanguage.INSTANCE),
+            "2.0.0"
+        )
+    }
+
+    fun `test (json) - 2_1_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.1.0.json", Language.findLanguageByID("JSON")!!),
+            "2.1.0"
+        )
+    }
+
+    fun `test (yaml) - 2_1_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.1.0.yaml", YAMLLanguage.INSTANCE),
+            "2.1.0"
+        )
+    }
+
+    fun `test (json) - 2_2_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.2.0.json", Language.findLanguageByID("JSON")!!),
+            "2.2.0"
+        )
+    }
+
+    fun `test (yaml) - 2_2_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.2.0.yaml", YAMLLanguage.INSTANCE),
+            "2.2.0"
+        )
+    }
+
+    fun `test (json) - 2_3_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.3.0.json", Language.findLanguageByID("JSON")!!),
+            "2.3.0"
+        )
+    }
+
+    fun `test (yaml) - 2_3_0 isSchema(psiFile) when file is schema and not empty`() {
+        `test - isSchema(psiFile) when file is schema and not empty`(
+            createFile("asyncapi-2.3.0.yaml", YAMLLanguage.INSTANCE),
+            "2.3.0"
+        )
+    }
+
+    private fun `test - isSchema(psiFile) when file is schema and not empty`(asyncApiFile: PsiFile, version: String) {
+        val asyncAPISchemaRecognizer = service<AsyncAPISchemaRecognizer>()
+
+        TestCase.assertTrue(
+            "$version must be supported.",
+            asyncAPISchemaRecognizer.isSupported(
+                asyncAPISchemaRecognizer.extractAsyncAPIKey(asyncApiFile)
+            )
+        )
+        TestCase.assertEquals(
+            "version $version must be recognized correctly",
+            version,
+            asyncAPISchemaRecognizer.extractAsyncAPIKey(asyncApiFile)
+        )
         TestCase.assertTrue(
             "must be true in case of non null, non empty json file with `asyncapi` key provided.",
             asyncAPISchemaRecognizer.isSchema(
-                (psiFileFactory.createFileFromText(
-                    "asyncapi.json",
-                    Language.findLanguageByID("JSON")!!,
-                    this.javaClass.getResource("/asyncapi.json")?.readText() ?: ""
-                ) as JsonFile)
-            )
-        )
-        TestCase.assertTrue(
-            "must be true in case of non null, non empty yaml file with `asyncapi` key provided.",
-            asyncAPISchemaRecognizer.isSchema(
-                (psiFileFactory.createFileFromText(
-                    "asyncapi.yaml",
-                    YAMLLanguage.INSTANCE,
-                    this.javaClass.getResource("/asyncapi.yaml")?.readText() ?: ""
-                ) as YAMLFile)
+                asyncApiFile
             )
         )
     }
 
-    fun `test - isSchema(psiFile) when file is empty`() {
+    fun `test (json) - 2_0_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.0.0.json", Language.findLanguageByID("JSON")!!, true)
+        )
+    }
+
+    fun `test (yaml) - 2_0_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.0.0.yaml", YAMLLanguage.INSTANCE, true)
+        )
+    }
+
+    fun `test (json) - 2_1_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.1.0.json", Language.findLanguageByID("JSON")!!, true)
+        )
+    }
+
+    fun `test (yaml) - 2_1_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.1.0.yaml", YAMLLanguage.INSTANCE, true)
+        )
+    }
+
+    fun `test (json) - 2_2_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.2.0.json", Language.findLanguageByID("JSON")!!, true)
+        )
+    }
+
+    fun `test (yaml) - 2_2_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.2.0.yaml", YAMLLanguage.INSTANCE, true)
+        )
+    }
+
+    fun `test (json) - 2_3_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.3.0.json", Language.findLanguageByID("JSON")!!, true)
+        )
+    }
+
+    fun `test (yaml) - 2_3_0 isSchema(psiFile) when file is empty`() {
+        `test - isSchema(psiFile) when file is empty`(
+            createFile("asyncapi-2.3.0.yaml", YAMLLanguage.INSTANCE, true)
+        )
+    }
+
+    private fun `test - isSchema(psiFile) when file is empty`(asyncApiFile: PsiFile) {
         val asyncAPISchemaRecognizer = service<AsyncAPISchemaRecognizer>()
 
+        TestCase.assertFalse(
+            "empty file must not be supported.",
+            asyncAPISchemaRecognizer.isSupported(
+                asyncAPISchemaRecognizer.extractAsyncAPIKey(asyncApiFile)
+            )
+        )
+        TestCase.assertNull(
+            "empty file must not be recognized correctly",
+            asyncAPISchemaRecognizer.extractAsyncAPIKey(asyncApiFile)
+        )
+        TestCase.assertFalse(
+            "must be false in case of empty file.",
+            asyncAPISchemaRecognizer.isSchema(asyncApiFile)
+        )
+    }
+
+    private fun createFile(asyncApi: String, language: Language, isEmpty: Boolean = false): PsiFile {
         val psiFileFactory = PsiFileFactory.getInstance(project)
-        TestCase.assertFalse(
-            "must be true in case of non null, non empty json file with `asyncapi` key provided.",
-            asyncAPISchemaRecognizer.isSchema(
-                (psiFileFactory.createFileFromText(
-                    "asyncapi.json",
-                    Language.findLanguageByID("JSON")!!,
-                    ""
-                ) as JsonFile)
-            )
-        )
-        TestCase.assertFalse(
-            "must be true in case of non null, non empty yaml file with `asyncapi` key provided.",
-            asyncAPISchemaRecognizer.isSchema(
-                (psiFileFactory.createFileFromText(
-                    "asyncapi.yaml",
-                    YAMLLanguage.INSTANCE,
-                    ""
-                ) as YAMLFile)
-            )
-        )
+
+        val text = if (isEmpty) {
+            ""
+        } else {
+            this.javaClass.getResource("/$asyncApi")?.readText() ?: ""
+        }
+
+        return psiFileFactory.createFileFromText(asyncApi, language, text)
     }
 
 }